+++ src/controllers/ProfileResource.php
@@ -2,11 +2,27 @@
 
 namespace roaresearch\yii2\roa\controllers;
 
-use roaresearch\yii2\roa\actions\{ProfileUpdate, ProfileView};
-use yii\{filters\VerbFilter, rest\OptionsAction};
+use roaresearch\yii2\roa\actions\{
+    ProfileFileStream,
+    ProfileUpdate,
+    ProfileView
+};
+use Yii;
+use yii\{base\Model, filters\VerbFilter, rest\OptionsAction};
 
 class ProfileResource extends \yii\rest\Controller
 {
+    /**
+     * @inheritdoc
+     */
+    public $updateScenario = Model::SCENARIO_DEFAULT;
+
+    /**
+     * @var string[] array used in `actions\Update::fileAttributes`
+     * @see actions\LoadFileTrait::$fileAttributes
+     */
+    public $updateFileAttributes = [];
+
     /**
      * @inheritdoc
      */
@@ -43,8 +59,13 @@ class ProfileResource extends \yii\rest\Controller
     {
         return [
             'view' => ['class' => ProfileView::class],
-            'update' => ['class' => ProfileUpdate::class],
+            'update' => [
+                'class' => ProfileUpdate::class,
+                'scenario' => $this->updateScenario,
+                'fileAttributes' => $this->updateFileAttributes,
+            ],
             'options' => ['class' => OptionsAction::class],
+            'file-stream' => ['class' => ProfileFileStream::class],
         ];
     }
 }